+++ ShopperInsights/src/main/java/com/braintreepayments/api/shopperinsights/v2/ShopperInsightsClientV2.kt
@@ -12,6 +12,7 @@ import com.braintreepayments.api.shopperinsights.ShopperInsightsAnalytics.BUTTON
 import com.braintreepayments.api.shopperinsights.ShopperInsightsAnalytics.BUTTON_SELECTED
 import com.braintreepayments.api.shopperinsights.v2.internal.CreateCustomerSessionApi
 import com.braintreepayments.api.shopperinsights.v2.internal.UpdateCustomerSessionApi
+import com.braintreepayments.api.shopperinsights.v2.internal.GenerateCustomerRecommendationsApi
 
 /**
  * Use [ShopperInsightsClientV2] to optimize your checkout experience by prioritizing the customer’s preferred payment
@@ -30,6 +31,8 @@ class ShopperInsightsClientV2 internal constructor(
     private val braintreeClient: BraintreeClient,
     private val createCustomerSessionApi: CreateCustomerSessionApi = CreateCustomerSessionApi(braintreeClient),
     private val updateCustomerSessionApi: UpdateCustomerSessionApi = UpdateCustomerSessionApi(braintreeClient),
+    private val generateCustomerRecommendationsApi: GenerateCustomerRecommendationsApi =
+        GenerateCustomerRecommendationsApi(braintreeClient),
     private val deviceInspector: DeviceInspector = DeviceInspector(),
     lazyAnalyticsClient: Lazy<AnalyticsClient> = AnalyticsClient.lazyInstance
 ) {
@@ -73,6 +76,7 @@ class ShopperInsightsClientV2 internal constructor(
     }
 
     /**
+<<<<<<< HEAD
      * Updates an existing customer session.
      *
      * @param customerSessionRequest: a [CustomerSessionRequest] object containing the request parameters
@@ -90,9 +94,47 @@ class ShopperInsightsClientV2 internal constructor(
             when (result) {
                 is UpdateCustomerSessionApi.UpdateCustomerSessionResult.Success ->
                     customerSessionCallback(CustomerSessionResult.Success(result.sessionId))
+
                 is UpdateCustomerSessionApi.UpdateCustomerSessionResult.Error ->
                     customerSessionCallback(CustomerSessionResult.Failure(result.error))
             }
+        }
+    }
+
+    /**
+     * Generates customer payment option recommendations.
+     *
+     * @param customerSessionRequest Optional: a [CustomerSessionRequest] object containing the request parameters
+     * @param sessionId Optional: The shopper session ID
+     * @param customerRecommendationsCallback: a callback that returns the result of the
+     * customer recommendation generation
+     *
+     * Note: **This feature is in beta. It's public API may change in future releases.**
+     */
+    fun generateCustomerRecommendations(
+        customerSessionRequest: CustomerSessionRequest? = null,
+        sessionId: String? = null,
+        customerRecommendationsCallback: (customerRecommendationsResult: CustomerRecommendationsResult) -> Unit
+    ) {
+        generateCustomerRecommendationsApi.execute(customerSessionRequest, sessionId) {
+            generateCustomerRecommendationsResult ->
+                when (generateCustomerRecommendationsResult) {
+                    is GenerateCustomerRecommendationsApi.GenerateCustomerRecommendationsResult.Success -> {
+                        customerRecommendationsCallback(
+                            CustomerRecommendationsResult.Success(
+                                generateCustomerRecommendationsResult.customerRecommendations
+                            )
+                        )
+                    }
+
+                    is GenerateCustomerRecommendationsApi.GenerateCustomerRecommendationsResult.Error -> {
+                        customerRecommendationsCallback(
+                            CustomerRecommendationsResult.Failure(
+                                generateCustomerRecommendationsResult.error
+                            )
+                        )
+                    }
+                }
         }
     }